@@ -1,24 +1,208 @@
-import React from "react";
-import { Link } from "react-router-dom";
+import React, { useState, forwardRef, useMemo } from "react";
+import {
+  Link as RouterLink,
+  LinkProps as RouterLinkProps,
+} from "react-router-dom";
+import Divider from "@mui/material/Divider";
+import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
+import MuiDrawer from "@mui/material/Drawer";
+import CssBaseline from "@mui/material/CssBaseline";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
+import IconButton from "@mui/material/IconButton";
+import MenuIcon from "@mui/icons-material/Menu";
+import List from "@mui/material/List";
+import ListItem from "@mui/material/ListItem";
+import ListItemButton from "@mui/material/ListItemButton";
+import ListItemIcon from "@mui/material/ListItemIcon";
+import ListItemText from "@mui/material/ListItemText";
+import PeopleIcon from "@mui/icons-material/People";
+import CasinoIcon from "@mui/icons-material/Casino";
+import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
+import ChevronRightIcon from "@mui/icons-material/ChevronRight";
+import { styled, useTheme, Theme, CSSObject } from "@mui/material/styles";
+/* TODO: Get FontAwesome working */
+/* import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"; */
+/* <FontAwesomeIcon icon="fa-solid fa-ranking-star" /> */
+
+const drawerWidth = 240;
+
+type ListItemLinkProps = {
+  primary: string;
+  to: string;
+  icon?: React.ReactElement;
+  open?: boolean;
+};
+
+const links: ListItemLinkProps[] = [
+  { primary: "Competition", to: "/competitions", icon: <CasinoIcon /> },
+  { primary: "Athlete", to: "/athletes", icon: <PeopleIcon /> },
+];
+
+const ListItemLink = (props: ListItemLinkProps) => {
+  const { icon, primary, to, open } = props;
+
+  const renderLink = useMemo(
+    () =>
+      forwardRef<HTMLAnchorElement, Omit<RouterLinkProps, "to">>(function Link(
+        itemProps,
+        ref
+      ) {
+        return <RouterLink to={to} ref={ref} {...itemProps} role={undefined} />;
+      }),
+    [to]
+  );
+
+  return (
+    <List>
+      <ListItem component={renderLink}>
+        <ListItemButton
+          sx={{
+            minHeight: 48,
+            justifyContent: open ? "initial" : "center",
+            px: 2.5,
+          }}
+        >
+          {icon ? (
+            <ListItemIcon
+              sx={{
+                minWidth: 0,
+                mr: open ? 3 : "auto",
+                justifyContent: "center",
+              }}
+            >
+              {icon}
+            </ListItemIcon>
+          ) : null}
+          <ListItemText sx={{ opacity: open ? 1 : 0 }} primary={primary} />
+        </ListItemButton>
+      </ListItem>
+    </List>
+  );
+};
+
+const openedMixin = (theme: Theme): CSSObject => ({
+  width: drawerWidth,
+  transition: theme.transitions.create("width", {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.enteringScreen,
+  }),
+  overflowX: "hidden",
+});
+
+const closedMixin = (theme: Theme): CSSObject => ({
+  transition: theme.transitions.create("width", {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.enteringScreen,
+  }),
+  overflowX: "hidden",
+  width: `calc(${theme.spacing(7)} + 1px)`,
+  [theme.breakpoints.up("sm")]: {
+    width: `calc(${theme.spacing(8)} + 1px)`,
+  },
+});
+
+const DrawerHeader = styled("div")(({ theme }) => ({
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "flex-end",
+  padding: theme.spacing(0, 1),
+  ...theme.mixins.toolbar,
+}));
+
+interface AppBarProps extends MuiAppBarProps {
+  open?: boolean;
+}
+
+const AppBar = styled(MuiAppBar, {
+  shouldForwardProp: (prop) => prop !== "open",
+})<AppBarProps>(({ theme, open }) => ({
+  zIndex: theme.zIndex.drawer + 1,
+  transition: theme.transitions.create(["width", "margin"], {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  ...(open && {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  }),
+}));
+
+const Drawer = styled(MuiDrawer, {
+  shouldForwardProp: (prop) => prop !== "open",
+})(({ theme, open }) => ({
+  width: drawerWidth,
+  flexShrink: 0,
+  whiteSpace: "nowrap",
+  boxSizing: "border-box",
+  ...(open && {
+    ...openedMixin(theme),
+    "& .MuiDrawer-paper": openedMixin(theme),
+  }),
+  ...(!open && {
+    ...closedMixin(theme),
+    "& .MuiDrawer-paper": closedMixin(theme),
+  }),
+}));
 
 const Navbar = () => {
+  const theme = useTheme();
+  const [open, setOpen] = useState(false);
+
+  const handleDrawerOpen = () => {
+    setOpen(true);
+  };
+
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
   return (
-    <div className="px-6 py-4 flex flex-rows border border-b-slate-400">
-      <div className="basis-1/6">
-        <Link to="/" className="hover:underline">
-          Lifters
-        </Link>
-      </div>
-      <div className="basis-4/6 flex gap-4">
-        <Link to="/competitions/" className="hover:underline">
-          Competitions
-        </Link>
-        <Link to="/athletes/" className="hover:underline">
-          Athletes
-        </Link>
-      </div>
-      <div className="basis-1/6"></div>
-    </div>
+    <>
+      <CssBaseline />
+      <AppBar position="fixed" open={open}>
+        <Toolbar>
+          <IconButton
+            color="inherit"
+            aria-label="open drawer"
+            onClick={handleDrawerOpen}
+            edge="start"
+            sx={{ marginRight: 5, ...(open && { display: { sm: "none" } }) }}
+          >
+            <MenuIcon />
+          </IconButton>
+          <Typography variant="h6" noWrap component="div">
+            Weightlifting Results for New Zealand
+          </Typography>
+        </Toolbar>
+      </AppBar>
+      <Drawer variant="permanent" open={open}>
+        <DrawerHeader>
+          <IconButton onClick={handleDrawerClose}>
+            {theme.direction === "rtl" ? (
+              <ChevronRightIcon />
+            ) : (
+              <ChevronLeftIcon />
+            )}
+          </IconButton>
+        </DrawerHeader>
+        <Divider />
+        {links.map((link, idx) => {
+          return (
+            <ListItemLink
+              key={idx}
+              icon={link.icon}
+              to={link.to}
+              primary={link.primary}
+              open={open}
+            />
+          );
+        })}
+      </Drawer>
+    </>
   );
 };
 
